+++ src/S3/Extensions/Context/CakeEnvironmentExtensions.cs
@@ -74,7 +74,7 @@ namespace Cake.AWS.S3
         /// Helper method to get the AWS Credentials from environment variables
         /// </summary>
         /// <param name="environment">The cake environment.</param>
-        /// <returns>A new <see cref="DownloadSettings"/> instance to be used in calls to the <see cref="IS3Manager"/>.</returns>
+        /// <returns>A new <see cref="UploadSettings"/> instance to be used in calls to the <see cref="IS3Manager"/>.</returns>
         public static UploadSettings CreateUploadSettings(this ICakeEnvironment environment)
         {
             if (environment == null)
@@ -83,6 +83,21 @@ namespace Cake.AWS.S3
             }
 
             return environment.SetSettings(new UploadSettings());
+        }
+        
+        /// <summary>
+        /// Helper method to get the AWS Credentials from environment variables
+        /// </summary>
+        /// <param name="environment">The cake environment.</param>
+        /// <returns>A new <see cref="SyncSettings"/> instance to be used in calls to the <see cref="IS3Manager"/>.</returns>
+        public static SyncSettings CreateSyncSettings(this ICakeEnvironment environment)
+        {
+            if (environment == null)
+            {
+                throw new ArgumentNullException("environment");
+            }
+
+            return environment.SetSettings(new SyncSettings());
         }
     }
 }